SDKQE-2769: correctly clean up dns records
For SRV records, when there are multiple clusters, the output of list records will contain all of the insecure records followed by all of the secure records rather than grouping the records for each cluster together. Fix the existence check by checking for each record separately. The same issue is present for A records but the hostnames for node1 of each cluster will be followed by node2 etc..